Fix error printing
This is why we need clap 3. Use eprint() instead of eprintln(), because
apparently the help & version text already contains the newline. Might
need to match on the error kind for other types of errors to print the
newline, but we'll have to see if that ever comes up.

Checked with a diff tool that the formatting of the error texts is
correct (in fact better than before!).