Fix behavior of search mode change funcitons
Revert to calling advance_search("") in search mode changes. This is a bit hacky as mentioned, but it makes more sense to do this inside the methods of the app state than having to do it in the UI. This is towards the ideal that the app state should always be consistent.